adjust ApolloClientProperties
