+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/extension/properties/ApolloClientExtensionProperties.java
@@ -16,6 +16,9 @@
  */
 package com.ctrip.framework.apollo.config.data.extension.properties;
 
+import com.ctrip.framework.apollo.config.data.extension.enums.ApolloClientMessagingType;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+
 /**
  * @author vdisk <vdisk@foxmail.com>
  */
@@ -26,6 +29,17 @@ public class ApolloClientExtensionProperties {
    */
   private Boolean enabled = false;
 
+  /**
+   * apollo client listening type
+   */
+  private ApolloClientMessagingType messagingType = ApolloClientMessagingType.LONG_POLLING;
+
+  /**
+   * apollo client authentication properties
+   */
+  @NestedConfigurationProperty
+  private ApolloClientAuthenticationProperties authentication;
+
   public Boolean getEnabled() {
     return enabled;
   }
@@ -34,10 +48,30 @@ public class ApolloClientExtensionProperties {
     this.enabled = enabled;
   }
 
+  public ApolloClientMessagingType getMessagingType() {
+    return messagingType;
+  }
+
+  public void setMessagingType(
+      ApolloClientMessagingType messagingType) {
+    this.messagingType = messagingType;
+  }
+
+  public ApolloClientAuthenticationProperties getAuthentication() {
+    return authentication;
+  }
+
+  public void setAuthentication(
+      ApolloClientAuthenticationProperties authentication) {
+    this.authentication = authentication;
+  }
+
   @Override
   public String toString() {
     return "ApolloClientExtensionProperties{" +
         "enabled=" + enabled +
+        ", messagingType=" + messagingType +
+        ", authentication=" + authentication +
         '}';
   }
 }
