+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/extension/properties/ApolloClientProperties.java
@@ -16,7 +16,6 @@
  */
 package com.ctrip.framework.apollo.config.data.extension.properties;
 
-import com.ctrip.framework.apollo.config.data.extension.enums.ApolloClientMessagingType;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
 
 /**
@@ -25,30 +24,10 @@ import org.springframework.boot.context.properties.NestedConfigurationProperty;
 public class ApolloClientProperties {
 
   /**
-   * apollo client listening type
-   */
-  private ApolloClientMessagingType messagingType = ApolloClientMessagingType.LONG_POLLING;
-
-  /**
    * extension configuration
    */
   @NestedConfigurationProperty
   private ApolloClientExtensionProperties extension;
-
-  /**
-   * apollo client authentication properties
-   */
-  @NestedConfigurationProperty
-  private ApolloClientAuthenticationProperties authentication;
-
-  public ApolloClientMessagingType getMessagingType() {
-    return messagingType;
-  }
-
-  public void setMessagingType(
-      ApolloClientMessagingType messagingType) {
-    this.messagingType = messagingType;
-  }
 
   public ApolloClientExtensionProperties getExtension() {
     return extension;
@@ -59,21 +38,10 @@ public class ApolloClientProperties {
     this.extension = extension;
   }
 
-  public ApolloClientAuthenticationProperties getAuthentication() {
-    return authentication;
-  }
-
-  public void setAuthentication(
-      ApolloClientAuthenticationProperties authentication) {
-    this.authentication = authentication;
-  }
-
   @Override
   public String toString() {
     return "ApolloClientProperties{" +
-        "messagingType=" + messagingType +
-        ", extension=" + extension +
-        ", authentication=" + authentication +
+        "extension=" + extension +
         '}';
   }
 }
